fix(harness-init): address review — symlink-safe probe, test split, dismissal logging
Address CodeRabbit review on #5066 (#5047):
- security(runtime_python): reject symlinked cache entries in
  probe_any_managed_install before the is_dir() check. path.is_dir()
  follows symlinks, so a link under cache_root could point the durable
  probe at a dir outside the root and be trusted as managed Python;
  entry.file_type() reports the link itself and is skipped. (The Node
  bootstrap already gets equivalent protection from its canonicalize +
  starts_with(cache_root) guard.)
- refactor(runtime_node): extract the inline probe_installed_tests module
  into runtime_node/bootstrap_tests.rs, wired via
  #[cfg(test)] #[path = "bootstrap_tests.rs"] mod tests; — matching the
  sibling runtime_python pattern and taking bootstrap.rs back under the
  ~500-line ceiling.
- logging(HarnessInitOverlay): add namespaced harness-init debug
  diagnostics across the dismissal-persistence flow — read/write helpers
  (incl. sessionStorage-failure paths), the warm-poll stay-hidden branch,
  and the handleContinue dismissal action.

Targeted: cargo fmt; cargo test probe_installed (6 passed, incl. the moved
node module); prettier on the overlay.
